@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 import axios from 'axios'
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 const AddUser = () => {
-  let history = useHistory();
+  let history = useNavigate();
   const [user, setUser] = useState({
     name: "",
     username: "",
